refactor: domain-> entity 엔티티들을 더 작은 도메인으로 쪼개기 위해 패키지명 수정
